TT_Backend/requirements.txt
asgiref==3.8.1
Django==4.1.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
mongoengine==0.28.2
PyJWT==2.9.0
pymongo==3.12.1
pytz==2024.1
sqlparse==0.2.4
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
uuid==1.30
